requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
numpy>=1.20.0
matplotlib>=3.4.0
Pillow>=8.2.0
aiohttp>=3.7.4
async-timeout>=3.0.1
python-dateutil>=2.8.2
